Move boost and tag into another file
