+++ fedireads/activitypub/__init__.py
@@ -4,12 +4,10 @@ import sys
 
 from .base_activity import ActivityEncoder, Image, PublicKey, Signature
 from .note import Note, Article, Comment, Review, Quotation
-from .note import Boost, Like
+from .interaction import Boost, Like
 from .ordered_collection import OrderedCollection, OrderedCollectionPage
 from .person import Person
-
 from .book import Edition, Work, Author
-
 from .verbs import Create, Undo, Update
 from .verbs import Follow, Accept, Reject
 from .verbs import Add, Remove
